Split api/client/membership.rs into directory
